+++ src/PSRule/Pipeline/PipelineWriterExtensions.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using PSRule.Resources;
+
+namespace PSRule.Pipeline
+{
+    internal static class PipelineWriterExtensions
+    {
+        internal static void WarnUsingInvariantCulture(this PipelineWriter writer)
+        {
+            if (!writer.ShouldWriteWarning())
+                return;
+
+            writer.WriteWarning(PSRuleResources.UsingInvariantCulture);
+        }
+    }
+}